requirements.txt
flask
line-bot-sdk
azure-cognitiveservices-vision-face
msrestazure